requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
qrcode[pil]==7.4.2
pyzbar==0.1.9
Pillow==10.1.0
sqlalchemy==2.0.23
pytest==7.4.3
httpx==0.25.2
